+++ app/src/main/java/com/wannagohome/lens_review_android/network/model/review/ReviewPreview.kt
@@ -41,31 +41,4 @@ data class ReviewPreview(
     val lens: LensPreview
 
 
-) {
-
-    fun getDateTime(): String {
-
-
-        return if (Build.VERSION_CODES.O <= Build.VERSION.SDK_INT) {
-            val dateTime = ZonedDateTime.parse(createdAt)
-                .withZoneSameInstant(ZoneId.systemDefault())
-
-            val month = dateTime.monthValue
-            val date = dateTime.dayOfMonth
-            val hours = dateTime.hour
-            val minutes = dateTime.minute
-
-            "${month}/$date-$hours:$minutes"
-        } else {
-            val dateTime = org.threeten.bp.ZonedDateTime.parse(createdAt)
-                .withZoneSameInstant(org.threeten.bp.ZoneId.systemDefault())
-
-            val month = dateTime.monthValue
-            val date = dateTime.dayOfMonth
-            val hours = dateTime.hour
-            val minutes = dateTime.minute
-
-            "${month}/$date-$hours:$minutes"
-        }
-    }
-}
+)
